Product Review Aggregator/requirements.txt
Django==5.0.2
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.12.0
python-dotenv==1.0.1
whitenoise==6.6.0
google-generativeai==0.3.2 